test: added .py script to generate the SQL inserts based on the .csv

+++ app/database/create_tables.sql
@@ -14,11 +14,12 @@ WHERE deleted_at IS NULL;
 
 CREATE TABLE IF NOT EXISTS user_emails (
     id BIGSERIAL PRIMARY KEY,
-    user_id INT NOT NULL,
+    user_id BIGINT NOT NULL,
     email VARCHAR(100) NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     verified_at TIMESTAMP,
     deleted_at TIMESTAMP,
+
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 );
 
@@ -31,12 +32,13 @@ WHERE deleted_at IS NULL;
 
 CREATE TABLE IF NOT EXISTS user_email_verification_tokens (
     id BIGSERIAL PRIMARY KEY,
-    user_email_id INT NOT NULL,
+    user_email_id BIGINT NOT NULL,
     token VARCHAR(255) NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     expires_at TIMESTAMP NOT NULL,
     deleted_at TIMESTAMP,
     verified_at TIMESTAMP,
+
     FOREIGN KEY (user_email_id) REFERENCES user_emails(id) ON DELETE CASCADE
 );
 
@@ -46,10 +48,11 @@ WHERE deleted_at IS NULL AND verified_at IS NULL AND expires_at > CURRENT_TIMEST
 
 CREATE TABLE IF NOT EXISTS user_password_hashes (
     id BIGSERIAL PRIMARY KEY,
-    user_id INT NOT NULL,
+    user_id BIGINT NOT NULL,
     password_hash VARCHAR(255) NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     deleted_at TIMESTAMP,
+
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 );
 
@@ -59,12 +62,13 @@ WHERE deleted_at IS NULL;
 
 CREATE TABLE IF NOT EXISTS user_password_reset_tokens (
     id BIGSERIAL PRIMARY KEY,
-    user_id INT NOT NULL,
+    user_id BIGINT NOT NULL,
     token VARCHAR(255) NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     expires_at TIMESTAMP NOT NULL,
     deleted_at TIMESTAMP,
     used_at TIMESTAMP,
+
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 );
 
@@ -74,12 +78,13 @@ WHERE deleted_at IS NULL AND used_at IS NULL AND expires_at > CURRENT_TIMESTAMP;
 
 CREATE TABLE IF NOT EXISTS user_refresh_tokens (
     id BIGSERIAL PRIMARY KEY,
-    user_id INT NOT NULL,
+    user_id BIGINT NOT NULL,
     token VARCHAR(255) NOT NULL,
     parent_token_id BIGINT,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     expires_at TIMESTAMP NOT NULL,
     deleted_at TIMESTAMP,
+
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE,
     FOREIGN KEY (parent_token_id) REFERENCES user_refresh_tokens(id) ON DELETE CASCADE
 );
@@ -90,7 +95,7 @@ WHERE parent_token_id IS NOT NULL;
 
 CREATE TABLE IF NOT EXISTS user_access_tokens (
     id BIGSERIAL PRIMARY KEY,
-    user_id INT NOT NULL,
+    user_id BIGINT NOT NULL,
     token VARCHAR(255) NOT NULL,
     parent_token_id BIGINT,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -114,6 +119,7 @@ CREATE TABLE IF NOT EXISTS jokes (
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     deleted_at TIMESTAMP,
+
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 );
 
@@ -144,3 +150,4 @@ CREATE TABLE IF NOT EXISTS joke_comments (
     FOREIGN KEY (joke_id) REFERENCES jokes(id) ON DELETE CASCADE,
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 );
+
